[GameRunner2 Extract Methods] Move static utility methods that were moved into ClientContext, mover their implementations to a new class 'ClientFileSystemHelper'
